My Weather App
After 3 month of intensiv study oh javascript, this app is what I am able to produce. It uses geolocation to display the weather on user location. It asks for location to be enabled. It takes a city written in english. 
The api service is from api.openweathermap.org.
I know is not the most beautiful Weather app out there, but for me is an achievement so far!